+++ zone/lua_client.h
@@ -493,8 +493,8 @@ public:
 	void SummonItemIntoInventory(luabind::object item_table);
 	bool HasItemOnCorpse(uint32 item_id);
 	void ClearXTargets();
-	uint8 GetAAEXPPercentage();
-	uint8 GetEXPPercentage();
+	int GetAAEXPPercentage();
+	int GetEXPPercentage();
 
 	void ApplySpell(int spell_id);
 	void ApplySpell(int spell_id, int duration);
